Moved .internal.hayward contents to .internal
